fix routes彙整成api.ts的錯誤
