+++ backend/src/app.ts
@@ -1,38 +1,21 @@
-import pathRoutes from './routes/api';
-import userRoutes from './routes/api';
-import testRoutes from './routes/testRoutes';
-import filesRoutes from './routes/api';
-import hostMsgRoutes from './routes/api';
-import compressRoutes from './routes/api';
 import express from 'express';
 import cors from 'cors';
 import './config/database'; // 執行`自執行函數`
+import apiRoutes from './routes/api'
+import { errorHandler } from './middlewares/errorHandler';
 
 const app = express();
 
 // 允許所有來源訪問 API
 // app.use(cors());
 // 嚴格限制來源訪問
-app.use(cors({
-  origin: 'http://localhost:3001'
-}));
+app.use(cors({ origin: 'http://localhost:3001' }));
  // 讓 Express 支援 JSON 請求體
 app.use(express.json());
 
-// for test
-app.use('/api/users', userRoutes);
-app.use('/api/test', testRoutes);
+app.use('/api', apiRoutes);
 
-// update & select path
-app.use('/api/path', pathRoutes);
-
-// view target file directory
-app.use('/api/files', filesRoutes);
-
-// compress file to zip
-app.use('/api/compress', compressRoutes);
-
-// host message
-app.use('/api/hostMsg', hostMsgRoutes);
+// 所有路由註冊之後
+app.use(errorHandler);
 
 export default app;
